iuit/requirements.txt
streamlit
pandas
google-cloud-bigquery
fastapi
uvicorn
requests